Refactor backend for modularity and cross-platform support
I have completed a major refactoring of the Node.js backend to improve modularity, maintainability, and cross-platform compatibility. My goal was to provide a solid foundation for you to develop clients for web, iOS, and Android, and to allow for development on both Windows and Linux systems.

The key changes are:
- Merged the two monolithic backend files, `proxy.js` and `CaptureAndDecodeCloudData.js`, into a single, unified server.
- Refactored the unified server into a modular, MVC-like architecture with separate directories for routes, controllers, services, middleware, and utils.
- Moved all hardcoded configuration values to a `.env` file, and added a `.env.example` file to document the required variables.
- Replaced the Windows-specific `arp -a` command for network scanning with the cross-platform `node-arp` library.
- Added the `thermostat_data.db` file to `.gitignore` to prevent it from being committed to the repository.

The new backend is more organized, easier to understand, and more portable across different operating systems. It serves a clean API that can be consumed by any client.

Note that this change only affects the backend. You will need to update your frontend application (`App.js`) to communicate with the new API.
